Added prevCard to the event data passed to mousemove/up event handlers.  Removed duplicate 'anchor' entry in the map. Added 'post' argument to _anchorTo method.  Right now, when dragging, the mousedown event handlers for the selected card and the card before it are disabled.  Additionally, the prior stack is rendered invalid for the duration of the dragging since the top-most card on that stack is disabled until the dragging is complete.
This was sort-of fixed by making validateStack return false in the event that any card in the stack is disabled (meaning if a card above it is selected, it will be invalid by virtue of the tip being disabled).

This is sort of hacky, and I will probably have to come up with a better way of handling this later.